requirements.txt
fastapi==0.121.2
uvicorn==0.34.0
pydantic==2.10.5
starlette==0.49.3
numpy==1.26.4
sentence-transformers==5.1.2
faiss-cpu==1.12.0
PyMuPDF==1.25.2
google-generativeai==0.8.5